Change amount from integer to float

+++ foundations-to-practice/version-1/part-1/schema.sql
@@ -17,6 +17,6 @@ CREATE TABLE recipe_ingredients (
   id SERIAL PRIMARY KEY,
   recipe_id INTEGER REFERENCES recipes,
   ingredient_id INTEGER REFERENCES ingredients,
-  amount INTEGER,
+  amount FLOAT,
   amount_units VARCHAR(8)
 );
